Fixed arguments to callList from controllers
Also added hack to refresh contact list data to avoid
NULL problems when iterating through contacts' students.

+++ grails-app/controllers/org/bworks/bworksdb/ClassSessionController.groovy
@@ -22,7 +22,7 @@ class ClassSessionController {
     }
 
     def welcomeLetters = {
-        def classSessionInstance = ClassSession.get(params.id)
+        def classSessionInstance = ClassSession.get(params.id.toLong())
         def contactInstanceList = classSessionInstance.enrollments.collect {
             it.student.contact
         }.unique()
